refactor(#156): modify folder name
