+++ src/config/type-orm/typeorm.config.ts
@@ -10,14 +10,14 @@ import { User } from 'src/users/entities/user.entity';
 import { CategoryList } from 'src/category/entity/category-list.entity';
 import { BadgeList } from 'src/common/entity/badge-list.entity';
 import { UserBadge } from 'src/users/entities/user-badge.entity';
-import { MentorReview } from 'src/mentors/entities/mentor-review.entity';
+import { MentorReview } from 'src/mentor-reviews/entities/mentor-review.entity';
 import { UserIntro } from 'src/users/entities/user-intro.entity';
 import { MentorBoardLike } from 'src/boards/entities/mentor-board-like.entity';
 import { MentorBoardImage } from 'src/boards/entities/mentor-board-image.entity';
 import { UserRanking } from 'src/users/entities/user-ranking.entity';
 import { TotalCount } from 'src/total-count/entities/total-count.entity';
 import { DataSource } from 'typeorm';
-import { MentorReviewChecklist } from 'src/mentors/mentor-review-checklist/entities/mentor-review-checklist.entity';
+import { MentorReviewChecklist } from 'src/mentor-reviews/mentor-review-checklist/entities/mentor-review-checklist.entity';
 
 // .env 파일 로드
 dotenv.config();
