+++ src/mentor-reviews/controllers/mentor-reviews.controller.ts
@@ -18,7 +18,7 @@ import {
 import { GetUserId } from 'src/common/decorators/get-userId.decorator';
 import { SuccessResponseInterceptor } from 'src/common/interceptors/success-response.interceptor';
 import { ParsePositiveIntPipe } from 'src/common/pipes/parse-positive-int.pipe';
-import { MentorsService } from '../services/mentors.service';
+import { MentorReviewsService } from '../services/mentor-reviews.service';
 import { CreateMentorReviewRequestBodyDto } from '../dtos/create-mentor-review-request-body.dto';
 import { JwtAccessTokenGuard } from 'src/config/guards/jwt-access-token.guard';
 import { MentorReviewDto } from '../dtos/mentor-review.dto';
@@ -41,8 +41,8 @@ import { ApiPatchUpdateMentorReview } from '../swagger-decorators/patch-update-m
 @UseInterceptors(SuccessResponseInterceptor, ClassSerializerInterceptor)
 @ApiTags('mentors-reviews')
 @Controller('mentors/:mentorId/reviews')
-export class MentorsController {
-  constructor(private readonly mentorsService: MentorsService) {}
+export class MentorReviewsController {
+  constructor(private readonly mentorsService: MentorReviewsService) {}
 
   @UseGuards(JwtAccessTokenGuard)
   @ApiCreateMentorReview()
